+++ backend/api/serializers/task_profiling.py
@@ -42,12 +42,8 @@ class TaskProfilingSerializer(serializers.ModelSerializer):
         ]
 
     def get_task_duration(self, obj: TaskProfiling) -> Union[str, None]:
-        if obj.compute_task.start_date is not None and obj.compute_task.end_date is not None:
-            duration = obj.compute_task.end_date - obj.compute_task.start_date
+        if obj.creation_date is not None and obj.compute_task.end_date is not None:
+            duration = obj.compute_task.end_date - obj.creation_date
             return duration_microseconds(duration)
         else:
             return None
-
-    def create(self, data):
-        task_profiling, created = TaskProfiling.objects.update_or_create(**data)
-        return task_profiling