+++ backend/api/tests/views/conftest.py
@@ -8,6 +8,7 @@ from rest_framework import test
 
 from api.models import ComputeTask
 from api.tests import asset_factory as factory
+from api.tests.common import AuthenticatedBackendClient
 from api.tests.common import AuthenticatedClient
 
 _CHANNEL_NAME: Final[str] = "mychannel"
@@ -26,6 +27,17 @@ def authenticated_client() -> test.APIClient:
 
     client.get = functools.partial(client.get, **_EXTRA_HTTP_HEADERS)
     client.post = functools.partial(client.post, **_EXTRA_HTTP_HEADERS)
+
+    return client
+
+
+@pytest.fixture
+def authenticated_backend_client() -> test.APIClient:
+    client = AuthenticatedBackendClient()
+
+    client.get = functools.partial(client.get, **_EXTRA_HTTP_HEADERS)
+    client.post = functools.partial(client.post, **_EXTRA_HTTP_HEADERS)
+    client.put = functools.partial(client.put, **_EXTRA_HTTP_HEADERS)
 
     return client