+++ backend/api/views/task_profiling.py
@@ -2,8 +2,11 @@ from typing import Any
 
 import structlog
 from django.db.models.query import QuerySet
+from django.db.utils import IntegrityError
+from django.utils import timezone
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import mixins
+from rest_framework import status
 from rest_framework.authentication import BasicAuthentication
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
@@ -23,6 +26,7 @@ logger = structlog.get_logger(__name__)
 
 class TaskProfilingViewSet(
     mixins.CreateModelMixin,
+    mixins.UpdateModelMixin,
     mixins.RetrieveModelMixin,
     mixins.ListModelMixin,
     GenericViewSet,
@@ -35,6 +39,18 @@ class TaskProfilingViewSet(
 
     def get_queryset(self) -> QuerySet[TaskProfiling]:
         return TaskProfiling.objects.filter(compute_task__channel=get_channel_name(self.request))
+
+    def create(self, request: Request, *args: Any, **kwargs: Any):
+        try:
+            task_profiling = super().create(request, *args, **kwargs)
+        except IntegrityError:
+            data = {"detail": f"TaskProfiling with key {request.data['compute_task_key']} already exists"}
+            return Response(data, status=status.HTTP_409_CONFLICT)
+        return task_profiling
+
+    def perform_update(self, serializer):
+        kwargs = {"creation_date": timezone.now()}
+        return serializer.save(**kwargs)
 
 
 class TaskProfilingStepViewSet(mixins.CreateModelMixin, GenericViewSet):