+++ backend/substrapp/exceptions.py
@@ -1,3 +1,6 @@
+from typing import Optional
+
+
 class KubernetesError(Exception):
     pass
 
@@ -28,6 +31,15 @@ class ImageDeletionError(Exception):
 
 class OrganizationError(Exception):
     """An error occurred during the download of an asset from a organization"""
+
+
+class OrganizationHttpError(OrganizationError):
+    """An error occurred during the request to an organization"""
+
+    def __init__(self, url: str, status_code: Optional[int] = None):
+        self.url = url
+        self.status_code = status_code
+        super().__init__(f"URL: {url} returned status code {status_code or 'unknown'}")
 
 
 class IntegrityError(Exception):